Move Java only tests from org.openhab.core.tests to org.openhab.core
These tests only require Java and not OSGi.
